refactor(stops-summary): extract StopsSummary into its own module
`StopsSummary` was an internal component of `stop-browser-header.tsx`
with its own logger (`NearbyStopsSummary`) and helper functions whose
naming made it hard to locate by grep: the file was named for the
header, the component was named for the summary, and the logger
referred to "nearby stops" instead of either.

- Move `StopsSummary`, `formatRadius`, the summary text builder, and
  `formatStopsCountsDebugLog` into a new `src/components/stops-summary.tsx`.
- Rename the logger from `NearbyStopsSummary` to `StopsSummary` so it
  matches the component name.
- Rename the text builder from `getNearbyStopsSummaryText` to
  `buildStopsSummaryText` (the function applies conditional formatting
  based on `hasLoaded` / `omitEmptyStops` / `infoLevel`, which is
  closer to "build" than "get") and to drop the now-redundant
  `Nearby` prefix.

No behavior change.
